refactor(Layout): move styles
